Added the funtions to agent_windows.go and called it from agent.go
