+++ cmd/antrea-agent/options_windows.go
@@ -60,7 +60,3 @@ func (o *Options) checkUnsupportedFeatures() error {
 	}
 	return nil
 }
-
-
-
-
